refactor: 인증 수정
