+++ src/main/java/practice/buttersaltflour/auth/entity/CustomPrincipal.java
@@ -1,4 +1,4 @@
-package practice.buttersaltflour.member.controller.dto;
+package practice.buttersaltflour.auth.entity;
 
 import lombok.Getter;
