+++ src/main/java/practice/buttersaltflour/filter/FirebaseAuthFilter.java
@@ -17,7 +17,7 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import practice.buttersaltflour.filter.exception.FirebaseAuthFailureException;
 import practice.buttersaltflour.filter.exception.FirebaseAuthHeaderMissingException;
 import practice.buttersaltflour.filter.exception.FirebaseAuthInvalidFormatException;
-import practice.buttersaltflour.member.controller.dto.CustomPrincipal;
+import practice.buttersaltflour.auth.entity.CustomPrincipal;
 
 import java.io.IOException;
 import java.util.List;
