+++ src/main/java/practice/buttersaltflour/member/service/MemberService.java
@@ -1,14 +1,12 @@
 package practice.buttersaltflour.member.service;
 
-import com.google.firebase.auth.FirebaseToken;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import practice.buttersaltflour.member.controller.dto.CustomPrincipal;
+import practice.buttersaltflour.auth.entity.CustomPrincipal;
 import practice.buttersaltflour.member.controller.dto.MemberResponse;
 import practice.buttersaltflour.member.entity.Member;
 import practice.buttersaltflour.member.exception.AuthContextMissingException;
